alter  type d'un attribut d'une methode CategorieService.java

+++ src/main/java/club/service/CategorieService.java
@@ -25,7 +25,7 @@ public class CategorieService {
     }
 
     // Obtenir une catégorie
-    public Categorie getCategorieById (Integer id) {
+    public Categorie getCategorieById (long id) {
         return categorieDAO.findById(id);
     }
